Extract draw data and window from context header
